ensure file was written only once

+++ tf/40_ecs_cluster.tf
@@ -19,9 +19,11 @@ sudo mount -t nfs4 -o nfsvers=4.1,rsize=1048576,wsize=1048576,hard,timeo=600,ret
 sudo mount -t nfs4 -o nfsvers=4.1,rsize=1048576,wsize=1048576,hard,timeo=600,retrans=2 ${aws_efs_mount_target.teamcity_log_dir.dns_name}:/ /opt/teamcity_log_dir
 sudo service docker restart
 mkdir -p /opt/teamcity_data_dir/config
-echo "connectionProperties.user=${var.teamcity_db_user}\n" > /opt/teamcity_data_dir/config/database.properties
-echo "connectionProperties.password=${var.teamcity_db_pass}\n" >> /opt/teamcity_data_dir/config/database.properties
-echo "connectionUrl=jdbc\:mysql\://${aws_db_instance.teamcity.address}/${var.teamcity_db_name}" >> /opt/teamcity_data_dir/config/database.properties
+if [! -e /opt/teamcity_data_dir/config/database.properties]; then
+  echo "connectionProperties.user=${var.teamcity_db_user}\n" > /opt/teamcity_data_dir/config/database.properties
+  echo "connectionProperties.password=${var.teamcity_db_pass}\n" >> /opt/teamcity_data_dir/config/database.properties
+  echo "connectionUrl=jdbc\:mysql\://${aws_db_instance.teamcity.address}/${var.teamcity_db_name}" >> /opt/teamcity_data_dir/config/database.properties
+fi
 echo ECS_CLUSTER=${aws_ecs_cluster.teamcity.name} > /etc/ecs/ecs.config
 EOF
 }
